pyproject.toml
[project]
name = "poc-monitor-contrato"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.6.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
